refactor v2 trace map step
